+++ opencti-platform/opencti-graphql/src/modules/playbook/playbook-utils.ts
@@ -71,7 +71,7 @@ export const filterBundleElements = async (
   context: AuthContext,
   bundleElements: StixObject[],
   filters: string | undefined,
-) => {
+): Promise<StixObject[]> => {
   if (!filters || isEmptyField(filters)) return bundleElements;
   const jsonFilters = JSON.parse(filters);
   const filterResults = await Promise.all(
